py_src/requirements.txt
transformers==4.57.3
torch==2.9.1
torchvision==0.24.1
pandas==2.2.3
numpy==2.2.0
Pillow==11.0.0
matplotlib==3.10.0
seaborn==0.13.2
scikit-learn==1.8.0
tqdm==4.67.1
